Refactor API client imports and update button components
