+++ src/ui/User.tsx
@@ -1,11 +1,8 @@
-import useUser from "../featuresHook/useUser";
+import { motion } from "framer-motion";
 import ButtonIcon from "./ButtonIcon";
 import SettingsIcon from "./icons/settings";
-import { motion } from "framer-motion";
 
 function User({ setShowSettings }) {
-  const { isLoading, user } = useUser();
-
   function handleShowSettings(e) {
     e.stopPropagation();
     setShowSettings((sh) => !sh);
@@ -22,8 +19,7 @@ function User({ setShowSettings }) {
       : "Good Night"
   }`;
 
-  if (isLoading) return <p>Loading User</p>;
-  const { username } = user;
+  // if (isLoading) return <p>Loading User</p>;
   return (
     <motion.div
       initial={{ opacity: 0, y: -10 }}
@@ -32,7 +28,7 @@ function User({ setShowSettings }) {
       className="flex gap-3 text-white ml-auto text-base lg:text-2xl md:text-1xl w-fit"
     >
       <p>
-        {timeString}, {username || "User"}
+        {timeString}, {"User"}
       </p>
       <ButtonIcon onClick={handleShowSettings}>
         {" "}
